+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## [Unreleased]
 
-## [0.9.0] - 2021-11-04
+## [0.9.2] - 2021-11-09
 
-- Initial release
+- Added optional TailwindCSS property handler
 
 ## [0.9.1] - 2021-11-08
 
-- Implement property handling
+- Implemented property handling
+
+## [0.9.0] - 2021-11-04
+
+- Initial release